Please don't kill me
I made a lot of changes but I think this should work. It takes -10, 0,
and 10 as inputs and outputs a float from 0 to 1. The two motors are
then multiplied by the (output) and (1 - output) to make a counter
turning thing. We can test. we will need to tweak values for kp ki and
kd as well possibly swapping the 1-output and output